requirements.txt
audioread==3.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
fonttools==4.43.1
freetype-py==2.4.0
glcontext==2.5.0
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
matplotlib==3.7.0
moderngl==5.8.2
msgpack==1.0.7
multipledispatch==1.0.0
numba==0.58.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.8.0
PyAudio==0.2.13
pycparser==2.21
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrr==0.10.3
python-dateutil==2.8.2
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.0.7
